+++ src/common.ts
@@ -1,5 +1,5 @@
 import dotenv from "dotenv";
-import { Collection, MongoClient } from "mongodb";
+import { MongoClient } from "mongodb";
 
 dotenv.config();
 
@@ -78,9 +78,8 @@ export function allequal(arr: any[]) {
 export const db = client.db("metahkg");
 
 export const threadCl = db.collection("thread");
-export const usersCl: Collection = db.collection("users");
+export const usersCl = db.collection("users");
 export const limitCl = db.collection("limit");
-export const imagesCl = db.collection("images");
 export const verificationCl = db.collection("verification");
 export const categoryCl = db.collection("category");
 export const votesCl = db.collection("votes");