+++ src/routes/categories.ts
@@ -1,13 +1,13 @@
-import profile from "./profile/profile";
-import avatars from "./profile/avatars";
 import { FastifyInstance, FastifyPluginOptions } from "fastify";
+import { categoryCl } from "../common";
 
 export default function (
     fastify: FastifyInstance,
     opts: FastifyPluginOptions,
     done: (e?: Error) => void
 ) {
-    fastify.register(profile);
-    fastify.register(avatars);
+    fastify.get("/", async (req, res) => {
+        res.send(await categoryCl.find().project({ _id: 0 }).sort({ id: 1 }).toArray());
+    });
     done();
 }