+++ src/routes/index.ts
@@ -1,9 +1,10 @@
-import users from "./routes/users";
-import category from "./routes/category";
-import menu from "./routes/menu";
-import thread from "./routes/thread";
-import profile from "./routes/profile";
+import users from "./users";
+import menu from "./";
+import thread from "./";
 import { FastifyInstance, FastifyPluginOptions } from "fastify";
+import me from "./me";
+import categories from "./categories";
+import category from "./category";
 
 export default (
     fastify: FastifyInstance,
@@ -11,9 +12,10 @@ export default (
     done: (e?: Error) => void
 ) => {
     fastify.register(users, { prefix: "/users" });
+    fastify.register(me, { prefix: "/me" });
+    fastify.register(categories, { prefix: "/categories" });
     fastify.register(category, { prefix: "/category" });
     fastify.register(menu, { prefix: "/menu" });
     fastify.register(thread, { prefix: "/thread" });
-    fastify.register(profile, { prefix: "/profile" });
     done();
 };