+++ src/routes/thread/comment/images.ts
@@ -1,9 +1,9 @@
 import isInteger from "is-sn-integer";
-import { threadCl } from "../../common";
+import { threadCl } from "../../../common";
 import { FastifyInstance, FastifyPluginOptions, FastifyRequest } from "fastify";
 import { Static, Type } from "@sinclair/typebox";
-import regex from "../../lib/regex";
-import Thread from "../../models/thread";
+import regex from "../../../lib/regex";
+import Thread from "../../../models/thread";
 
 export default (
     fastify: FastifyInstance,
@@ -12,10 +12,11 @@ export default (
 ) => {
     const paramsSchema = Type.Object({
         id: Type.RegEx(regex.integer),
+        cid: Type.RegEx(regex.integer),
     });
 
     fastify.get(
-        "/:id/images",
+        "/:id/comment/:cid/images",
         { schema: { params: paramsSchema } },
         async (
             req: FastifyRequest<{
@@ -26,11 +27,24 @@ export default (
             if (!isInteger(req.params.id))
                 return res.code(400).send({ error: "Bad request." });
 
-            const id = Number(req.params.id);
+            const threadId = Number(req.params.id);
+            const commentId = Number(req.params.cid);
 
             const result = (await threadCl.findOne(
-                { id },
-                { projection: { _id: 0, images: 1 } }
+                { id: threadId },
+                {
+                    projection: {
+                        _id: 0,
+                        images: {
+                            $filter: {
+                                input: "$images",
+                                cond: {
+                                    $eq: ["$$this.cid", commentId],
+                                },
+                            },
+                        },
+                    },
+                }
             )) as Thread;
 
             if (!result) return res.code(404).send({ error: "Not found." });